res_drivers_license: class definition moved for better overview
