+++ res_drivers_license/models/__init__.py
@@ -1,2 +1,1 @@
-from . import res_drivers_license
 from . import res_partner
