--- res_drivers_license/models/res_drivers_license.py
@@ -1,33 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    OpenERP, Open Source Management Solution, third party addon
-#    Copyright (C) 2004-2019 Vertel AB (<http://vertel.se>).
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-from odoo import models, fields, api, _
-import logging
-_logger = logging.getLogger(__name__)
-
-class ResDriversLicense(models.Model): 
-    _name = 'res.drivers_license'
-
-    partner_ids = fields.Many2many(comodel_name="res.partner")
-    
-    name = fields.Char(string='Class', required=True) #A,B etc.
-    description = fields.Char(string='Description')
-
-
